split async and sync codes for runc-shim and shim
