+++ crates/shim/examples/publish.rs
@@ -13,12 +13,13 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-
 use std::env;
 
-use containerd_shim::{Context, RemotePublisher};
+use containerd_shim::publisher::RemotePublisher;
+use containerd_shim::Context;
 use containerd_shim_protos::events::task::TaskOOM;
 
+#[cfg(not(feature = "async"))]
 fn main() {
     let args: Vec<String> = env::args().collect();
 
@@ -38,8 +39,39 @@ fn main() {
     let ctx = Context::default();
 
     println!("Sending event");
+
     publisher
         .publish(ctx, "/tasks/oom", "default", event)
+        .expect("Publish failed");
+
+    println!("Done");
+}
+
+#[cfg(feature = "async")]
+#[tokio::main]
+async fn main() {
+    let args: Vec<String> = env::args().collect();
+
+    // Must not start with unix://
+    let address = args
+        .get(1)
+        .ok_or("First argument must be containerd's TTRPC address to publish events")
+        .unwrap();
+
+    println!("Connecting: {}", &address);
+
+    let publisher = RemotePublisher::new(address).await.expect("Connect failed");
+
+    let mut event = TaskOOM::new();
+    event.set_container_id("123".into());
+
+    let ctx = Context::default();
+
+    println!("Sending event");
+
+    publisher
+        .publish(ctx, "/tasks/oom", "default", event)
+        .await
         .expect("Publish failed");
 
     println!("Done");
